extract useful functions to helpers module
